Update README.md
I have no experience in electronics like this but, those two resistors in your images that are smashed into the center of the breadboard look like 1000 Ohm resistors. The colors on them appear to be Brown Black *Red* Silver/Gold, when they should be Brown Black *Brown* Silver/Gold if they are 100 Ohm. I am basing this purely off of the images so disregard this if I am wrong. 

Does it really matter if they are 1000 Ohm or 100 Ohm?